pyproject.toml
[tool.black]
line-length = 88

[tool.poetry]
name = "church-class-list"
description = ""
version = "0.0.0"
authors = ["Trevor Lauder <trevor@lauderfamily.ca>"]

[tool.poetry.dependencies]
python = "^3.9"
pyyaml = "^5.4"
requests = "^2.26"

[tool.poetry.dev-dependencies]
black = "^21.7b0"
flake8 = "^3.9"
isort = "^5.9"
